requirements.txt
asgiref==3.2.7
boto3==1.13.10
botocore==1.16.10
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-crispy-forms==1.9.0
django-filter==2.2.0
django-six==1.0.4
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1